O basico do projecto
