+++ Makefile
@@ -6,18 +6,18 @@
 #    By: timatias <timatias@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/11/05 18:18:29 by timatias          #+#    #+#              #
-#    Updated: 2024/11/05 18:18:32 by timatias         ###   ########.fr        #
+#    Updated: 2024/11/05 19:56:41 by timatias         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 NAME = minishell
 
 CC = cc
-CFLAGS = -Wall -Wextra -Werror
+CFLAGS = -Wall -Wextra -Werror -Iinclude
 READLINE_FLAGS = -lreadline
 
 SRCS = src/main.c src/prompt.c src/parser.c src/executor.c \
-       src/builtins/echo.c src/builtins/cd.c src/builtins/pwd.c \
+       src/utils/echo.c src/utils/cd.c src/utils/pwd.c \
        src/utils/utils.c
 OBJS = $(SRCS:.c=.o)
