+++ src/executor.c
@@ -6,11 +6,11 @@
 /*   By: timatias <timatias@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 18:22:27 by timatias          #+#    #+#             */
-/*   Updated: 2024/11/05 18:23:05 by timatias         ###   ########.fr       */
+/*   Updated: 2024/11/05 18:55:23 by timatias         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../include/minishell.h"
 
 void    execute_command(char **args, t_shell *shell)
 {
