+++ src/parser.c
@@ -6,11 +6,11 @@
 /*   By: timatias <timatias@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 18:21:34 by timatias          #+#    #+#             */
-/*   Updated: 2024/11/05 18:22:01 by timatias         ###   ########.fr       */
+/*   Updated: 2024/11/05 20:03:53 by timatias         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../include/minishell.h"
 
 void parse_and_execute(char *input, t_shell *shell)
 {
@@ -18,12 +18,12 @@ void parse_and_execute(char *input, t_shell *shell)
     args[0] = input;
     args[1] = NULL;
 
-    if (strcmp(input, "exit") == 0)
+    if (ft_strcmp(input, "exit") == 0)
     {
         printf ("exit\n");
         exit (shell->last_exit_status);
     }
-    else if (strcmp(input, "echo") == 0)
+    else if (ft_strcmp(input, "echo") == 0)
         mini_echo(args);
     else
         printf("Comando não encontrado: %s\n", input);
