+++ src/utils/echo.c
@@ -6,7 +6,7 @@
 /*   By: timatias <timatias@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 18:23:45 by timatias          #+#    #+#             */
-/*   Updated: 2024/11/05 18:24:20 by timatias         ###   ########.fr       */
+/*   Updated: 2024/11/05 19:56:58 by timatias         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
