+++ src/utils/utils.c
@@ -1,23 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   cd.c                                               :+:      :+:    :+:   */
+/*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: timatias <timatias@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/05 18:24:05 by timatias          #+#    #+#             */
-/*   Updated: 2024/11/05 18:24:13 by timatias         ###   ########.fr       */
+/*   Created: 2024/11/05 20:11:26 by timatias          #+#    #+#             */
+/*   Updated: 2024/11/05 20:17:30 by timatias         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../include/minishell.h"
 
-int     mini_cd(char **args)
+int     ft_strcmp(const char *s1, const char *s2)
 {
-    if (args[1] == NULL || chdir(args[1]) != 0)
+    int     i;
+
+    i = 0;
+    while (s1[i] != '\0' && s2[i] != '\0' && (s1[i] == s2[i]))
     {
-        perror("minishell: cd");
-        return (1);
+        i++;
     }
-    return (0);
+    return (s1[i] - s2[i]);
 }
